+++ cart/src/main/java/com/syscho/wocom/carts/dto/CartItemDTO.java
@@ -1,5 +1,6 @@
-package com.syscho.wocom.carts;
+package com.syscho.wocom.carts.dto;
 
+import com.syscho.wocom.carts.products.ProductDTO;
 import lombok.Data;
 
 @Data
