+++ cart/src/main/java/com/syscho/wocom/carts/products/ProductDTO.java
@@ -1,4 +1,4 @@
-package com.syscho.wocom.carts;
+package com.syscho.wocom.carts.products;
 
 import lombok.Builder;
 import lombok.Data;
