+++ cart/src/main/java/com/syscho/wocom/carts/repo/CartRepository.java
@@ -1,4 +1,4 @@
-package com.syscho.wocom.carts;
+package com.syscho.wocom.carts.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
