+++ user/src/main/java/com/syscho/wocom/users/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package com.syscho.wocom.users.mapper;
 
-import com.syscho.wocom.users.UserDTO;
-import com.syscho.wocom.users.entity.UserEntity;
+import com.syscho.wocom.users.dto.UserDTO;
+import com.syscho.wocom.users.repo.UserEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
