+++ user/src/main/java/com/syscho/wocom/users/repo/UserRepository.java
@@ -1,4 +1,4 @@
-package com.syscho.wocom.users.entity;
+package com.syscho.wocom.users.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
